Split workspace automation engine
