+++ Tests/QuillCodeParityTests/ParityAutomationGateTests.swift
@@ -25,25 +25,43 @@ final class ParityAutomationGateTests: QuillCodeParityTestCase {
         let modelAutomationText = try Self.appSourceText(named: "WorkspaceModelAutomations.swift")
         let modelAutomationRunText = try Self.appSourceText(named: "WorkspaceModelAutomationRuns.swift")
         let automationText = try Self.appSourceText(named: "WorkspaceAutomationEngine.swift")
+        let automationFactoryText = try Self.appSourceText(named: "WorkspaceAutomationFactory.swift")
+        let automationReducerText = try Self.appSourceText(named: "WorkspaceAutomationStateReducer.swift")
+        let automationRunnerText = try Self.appSourceText(named: "WorkspaceAutomationRunner.swift")
 
         XCTAssertTrue(modelAutomationText.contains("extension QuillCodeWorkspaceModel"), "Automation model API should live in a focused workspace model extension.")
         XCTAssertTrue(modelAutomationRunText.contains("extension QuillCodeWorkspaceModel"), "Automation run API should live in a focused workspace model extension.")
-        XCTAssertTrue(automationText.contains("enum WorkspaceAutomationStateReducer"), "Automation state mutation should live in a focused reducer.")
-        XCTAssertTrue(automationText.contains("struct WorkspaceAutomationStateMutation"), "Automation state mutations should return typed mutation results.")
-        XCTAssertTrue(automationText.contains("static func setItems"), "Automation sorting and visibility preservation should be reducer-owned.")
-        XCTAssertTrue(automationText.contains("static func createThreadFollowUp"), "Thread follow-up creation should be reducer-owned.")
-        XCTAssertTrue(automationText.contains("static func createWorkspaceSchedule"), "Workspace schedule creation should be reducer-owned.")
-        XCTAssertTrue(automationText.contains("static func updateStatus"), "Automation status mutation should be reducer-owned.")
-        XCTAssertTrue(automationText.contains("static func delete("), "Automation deletion should be reducer-owned.")
-        XCTAssertTrue(automationText.contains("static func replace("), "Automation replacement should be reducer-owned.")
+        XCTAssertTrue(automationText.contains("public struct AutomationsState"), "Automation state records should stay in the focused data file.")
+        XCTAssertTrue(automationText.contains("public struct AutomationRunReport"), "Automation run report records should stay in the focused data file.")
+        XCTAssertTrue(automationText.contains("struct WorkspaceAutomationRunDraft"), "Automation run drafts should stay in the focused data file.")
+        XCTAssertTrue(automationText.contains("struct WorkspaceAutomationTrigger"), "Due automation execution should preserve trigger metadata.")
+        XCTAssertFalse(automationText.contains("enum WorkspaceAutomationFactory"), "Automation record construction should not drift into the data file.")
+        XCTAssertFalse(automationText.contains("enum WorkspaceAutomationStateReducer"), "Automation state mutation should not drift into the data file.")
+        XCTAssertFalse(automationText.contains("enum WorkspaceAutomationRunner"), "Automation run planning should not drift into the data file.")
+        XCTAssertTrue(automationFactoryText.contains("enum WorkspaceAutomationFactory"), "Automation construction should live in a focused factory.")
+        XCTAssertTrue(automationFactoryText.contains("static func threadFollowUp"), "Thread follow-up record construction should be factory-owned.")
+        XCTAssertTrue(automationFactoryText.contains("static func workspaceSchedule"), "Workspace schedule record construction should be factory-owned.")
+        XCTAssertTrue(automationFactoryText.contains("static func relativeSchedule"), "Relative schedule construction should be factory-owned.")
+        XCTAssertTrue(automationFactoryText.contains("static func tomorrowMorning"), "Default wall-clock schedule construction should be factory-owned.")
+        XCTAssertTrue(automationReducerText.contains("enum WorkspaceAutomationStateReducer"), "Automation state mutation should live in a focused reducer.")
+        XCTAssertTrue(automationReducerText.contains("struct WorkspaceAutomationStateMutation"), "Automation state mutations should return typed mutation results.")
+        XCTAssertTrue(automationReducerText.contains("static func setItems"), "Automation sorting and visibility preservation should be reducer-owned.")
+        XCTAssertTrue(automationReducerText.contains("static func createThreadFollowUp"), "Thread follow-up creation should be reducer-owned.")
+        XCTAssertTrue(automationReducerText.contains("static func createWorkspaceSchedule"), "Workspace schedule creation should be reducer-owned.")
+        XCTAssertTrue(automationReducerText.contains("static func updateStatus"), "Automation status mutation should be reducer-owned.")
+        XCTAssertTrue(automationReducerText.contains("static func delete("), "Automation deletion should be reducer-owned.")
+        XCTAssertTrue(automationReducerText.contains("static func replace("), "Automation replacement should be reducer-owned.")
         XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.setItems"), "WorkspaceModel automation extension should delegate automation item setting.")
         XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.createThreadFollowUp"), "WorkspaceModel automation extension should delegate thread follow-up creation.")
         XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.createWorkspaceSchedule"), "WorkspaceModel automation extension should delegate workspace schedule creation.")
         XCTAssertTrue(modelAutomationRunText.contains("WorkspaceAutomationStateReducer.updateStatus"), "WorkspaceModel automation run extension should delegate status changes.")
         XCTAssertTrue(modelAutomationRunText.contains("WorkspaceAutomationStateReducer.delete"), "WorkspaceModel automation run extension should delegate deletion.")
         XCTAssertTrue(modelAutomationRunText.contains("WorkspaceAutomationStateReducer.replace"), "WorkspaceModel automation run extension should delegate replacement.")
-        XCTAssertTrue(automationText.contains("struct WorkspaceAutomationTrigger"), "Due automation execution should preserve trigger metadata.")
-        XCTAssertTrue(automationText.contains("static func dueAutomationTriggers"), "Trigger-aware due selection should stay in the automation runner.")
+        XCTAssertTrue(automationRunnerText.contains("enum WorkspaceAutomationRunner"), "Automation run planning should live in a focused runner.")
+        XCTAssertTrue(automationRunnerText.contains("static func dueAutomationTriggers"), "Trigger-aware due selection should stay in the automation runner.")
+        XCTAssertTrue(automationRunnerText.contains("static func threadFollowUpDraft"), "Thread follow-up drafts should be runner-owned.")
+        XCTAssertTrue(automationRunnerText.contains("static func workspaceScheduleDraft"), "Workspace check drafts should be runner-owned.")
+        XCTAssertTrue(automationRunnerText.contains("static func monitorDraft"), "Monitor drafts should be runner-owned.")
         XCTAssertTrue(modelAutomationRunText.contains("automationEventSources()"), "WorkspaceModel automation run extension should resolve monitor event sources near run orchestration.")
         XCTAssertTrue(modelAutomationRunText.contains("eventDescription:"), "WorkspaceModel automation runs should pass event-source trigger descriptions into monitor drafts.")
         XCTAssertFalse(modelAutomationText.contains("WorkspaceAutomationStateReducer.updateStatus"), "WorkspaceModel automation creation extension should not own status changes.")
